Update network script name
There are two plays in deploy_undercloud.yaml that setup network
interfaces, one for eth1 and one for eth2. The `dest` argument of both
are the same, though. This causes the last ansible command to overwrite
the contents of the first.

This commit updates the second ansible play to write the eth1
configuration to ifcfg-eth1 instead of ifcfg-eth2.